@@ -1,5 +1,6 @@
-import EllipsisRasterLayer from './EllipsisRasterLayer';
-import EllipsisVectorLayer from './EllipsisVectorLayer';
-import { EllipsisApi } from 'ellipsis-js-util';
+import EllipsisRasterLayer from "./EllipsisRasterLayer";
+import EllipsisVectorLayer from "./EllipsisVectorLayer";
+
+import { EllipsisApi } from "ellipsis-js-util";
 
 export { EllipsisRasterLayer, EllipsisVectorLayer, EllipsisApi };
